Don't error when changing expense category (#49)
This fixes a production error that happened a few times:

```
Interview has an error.  There was a reference to a variable 'users[0].expenses[1].exists' that could not be looked up in the question file (for language 'en') or in any of the files incorporated by reference into the question file.
```

The issue is that when trying to review / edit an existing expense and
changing the category, DA decides to look at the `exists` attribute. The
root cause is that the `ALExpenseList` shouldn't have a
`complete_attribute` of exists anyway.

Additionally, we now correctly update changed category expense display
names, so it doesn't look like things aren't working to users.